pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ministudio"
version = "0.1.0"
description = "Model-Agnostic AI Video Generation Framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Ministudio Team", email = "team@ministudio.ai"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "gradio",
    "moviepy",
    "google-cloud-texttospeech",
    "boto3",
    "pillow",
    "numpy"
]

[project.optional-dependencies]
vertex-ai = ["google-genai"]
openai = ["openai", "requests"]
all = ["google-genai", "openai", "requests"]

[project.scripts]
ministudio = "ministudio.cli:main"

[tool.setuptools]
packages = ["ministudio"]
